+++ Sources/ACKLocalizationCore/Extensions/TokenExtensions.swift
@@ -1,0 +1,9 @@
+import Foundation
+import GoogleAuth
+
+extension Token: CredentialsType {
+    /// Adds `Authorization` header to given `request`
+    public func addToRequest(_ request: inout URLRequest) {
+        add(to: &request)
+    }
+}