+++ Tests/LinuxMain.swift
@@ -1,7 +1,0 @@
-import XCTest
-
-import ACKLocalizationTests
-
-var tests = [XCTestCaseEntry]()
-tests += ACKLocalizationTests.allTests()
-XCTMain(tests)